doc: third-party auth
