+++ api/schemas.py
@@ -1,5 +1,5 @@
-from ninja import ModelSchema
-from pydantic import Field
+from ninja import ModelSchema, Schema
+from pydantic import Field, HttpUrl
 
 from api.models import ApiClient
 from core.schemas import SimpleUserSchema
@@ -12,3 +12,12 @@ class ApiClientSchema(ModelSchema):
 
     owner: SimpleUserSchema
     permissions: list[str] = Field(alias="all_permissions")
+
+
+class ThirdPartyAuthParamsSchema(Schema):
+    client_id: int
+    third_party_app: str
+    privacy_link: HttpUrl
+    username: str
+    callback_url: HttpUrl
+    signature: str
