+++ docs/reference/api/schemas.md
@@ -1,0 +1,1 @@
+::: api.schemas
